requirements.txt
Django~=4.2.2
python-dotenv~=1.0.0
django-bootstrap-v5~=1.0.11
django-phonenumber-field~=7.1.0
pillow~=9.5.0
requests~=2.31.0
urllib3==1.26.7
django-allauth~=0.54.0
pytz~=2023.3
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
dj_rest_auth==2.2.0
drf_yasg==1.21.5